[SPARK-11734][SQL] Move reference sort into test and standardize on TungstenSort.
